detail style and delete auth

+++ client/public/scripts/cart/remove-item.js
@@ -9,11 +9,15 @@ document.addEventListener("DOMContentLoaded", () => {
     //REMOVE SELECTION
     removeItem.forEach(btn => {
         btn.addEventListener("click", () => {
-            let button = btn.dataset.id
-            let test = lsCart.filter(x => x.id != button)
-            localStorage.setItem("cart", JSON.stringify(test))
-            sendCart(test)
-            location.reload()
+            let res = window.confirm("Do you want to delete this item ?")
+            if (res) {
+                let button = btn.dataset.id
+                let test = lsCart.filter(x => x.id != button)
+                localStorage.setItem("cart", JSON.stringify(test))
+                sendCart(test)
+                location.reload()
+            }
+
         })
     })       
 })
